cleanup processor a bit
