+++ internal/eventstore/store.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"log"
 	"sync"
 	"time"
 
@@ -128,7 +127,7 @@ func (s *Store) Append(ctx context.Context, events ...Event) error {
 		if snap != nil {
 			rawData = snap.Data
 		}
-		data, err := p.Apply(rawData, events...)
+		data, err := p.RawApply(rawData, events)
 		if err != nil {
 			return err
 		}
@@ -176,61 +175,4 @@ func (s *Store) getSnapshot(ctx context.Context, tx PgConn, streamID string) (*S
 	}
 
 	return &snap, nil
-}
-
-type Processor interface {
-	Apply(json.RawMessage, ...Event) (any, error)
-}
-
-func Handler[T, TT any](fn func(T, TT) (T, error)) func(T, any) (T, error) {
-	return func(t T, a any) (T, error) {
-		return fn(t, a.(TT))
-	}
-}
-
-type ProcessorFor[T any] struct {
-	handlers   map[string]func(T, any) (T, error)
-	handlersMu sync.RWMutex
-}
-
-func (p *ProcessorFor[T]) AddHandler(eventType string, handler func(T, any) (T, error)) {
-	p.handlersMu.Lock()
-	defer p.handlersMu.Unlock()
-	if p.handlers == nil {
-		p.handlers = make(map[string]func(T, any) (T, error))
-	}
-	p.handlers[eventType] = handler
-}
-
-func (p *ProcessorFor[T]) Apply(d json.RawMessage, events ...Event) (any, error) {
-	var data T
-	if d != nil {
-		if err := json.Unmarshal(d, data); err != nil {
-			return nil, fmt.Errorf("eventstore: can't deserialize into %T: %w", data, err)
-		}
-	}
-
-	var err error
-
-	for _, e := range events {
-		log.Printf("eventstore: before applying event %+v", data)
-		data, err = p.applyEvent(data, e)
-		log.Printf("eventstore: after applying event %+v", data)
-		if err != nil {
-			return data, fmt.Errorf("eventstore: failed to apply %s event %s: %w", e.StreamType, e.Type, err)
-		}
-	}
-
-	return data, err
-}
-
-func (p *ProcessorFor[T]) applyEvent(data T, e Event) (T, error) {
-	p.handlersMu.RLock()
-	handler, ok := p.handlers[e.Type]
-	p.handlersMu.RUnlock()
-	if !ok {
-		return data, fmt.Errorf("eventstore: no handler for %s event %s", e.StreamType, e.Type)
-	}
-
-	return handler(data, e.Data)
 }
